tests/wpt/web-platform-tests/websockets/Create-valid-url-protocol-empty.htm

<!DOCTYPE html>
<html>
<head>
    <title>W3C WebSocket API - Create WebSocket - wsocket.protocol is empty before connection is established</title>
    <script type="text/javascript" src="/resources/testharness.js"></script>
    <script type="text/javascript" src="/resources/testharnessreport.js"></script>
    <script type="text/javascript" src="websocket.sub.js"></script>
</head>
<body>
    <div id="log"></div>
    <script type="text/javascript">
        test(function () {
            var wsocket = CreateWebSocket(false, true, false);
            assert_equals(wsocket.protocol, "", "protocol should be empty");
            wsocket.close();
        }, "W3C WebSocket API - Create WebSocket - wsocket.protocol should be empty before connection is established")
    </script>
</body>
</html>
